<fix>(pbft,txpool): refuse pre-prepare package which view is too larger than local view, to solve txs not unsealed bug. (#4308)
The scenario involves four nodes, with three nodes having a view of 10 and the fourth node having a view of 10000. Currently, there have been 100 blocks committed to consensus. When it's the turn of the node with a view of 10000 to propose a block, this node broadcasts a pre-prepare message for block 101. This proposal will eventually be rejected in later stages, but the cache has already stored this pre-prepare message. Additionally, even after a view change, this cache will not be cleared, and the transactions in the proposal will not be marked as unsealed, leading to a continuous backlog of transactions. Furthermore, after the view change, the pre-prepare message sent by the next leader (with a normal view) is considered conflicting with the pre-prepare message in the cache and is rejected. This continues until all nodes reach a view of 10000 before normal block proposal can resume.